+++ lib/pure.dart
@@ -1,22 +1,52 @@
 library pure;
 
-// -- Composition --
+// --- Common --- //
+
+export 'src/common/curried_function_types.dart';
+export 'src/common/function_types.dart';
+
+// -- Composition -- //
 
 export 'src/composition/dot.dart';
 export 'src/composition/pipe.dart';
 
-// -- Memoization --
+// --- Constant --- //
+
+export 'src/constant/extensions.dart';
+
+// --- Curry --- //
+
+export 'src/curry/extensions.dart';
+
+// --- Flip --- //
+
+export 'src/flip/extensions.dart';
+
+// -- Functions -- //
+
+export 'src/functions/functions.dart';
+
+// -- Memoization -- //
 
 export 'src/memoization/extensions.dart';
 
-// -- Partial Application --
+// -- Nullable -- //
+
+export 'src/nullable/extensions.dart';
+
+// -- Partial Application -- //
 
 export 'src/partial_application/extensions.dart';
 
-// -- Recursion Trampoline --
+// -- Recursion Trampoline -- //
 
-export 'src/recursion_trampoline/base.dart';
 export 'src/recursion_trampoline/extensions.dart';
+export 'src/recursion_trampoline/tram.dart';
+
+// -- Uncurrying -- //
+
+export 'src/uncurry/extensions.dart';
 
 // Convenience mixin for auto-imports
+
 mixin ImportPure {}