+++ CHANGELOG.md
@@ -1,7 +1,26 @@
-## [0.1.0] - 19.09.2021
+## [0.2.0] - 20.03.2022
 
-- Initial version.
+### Changed
+
+- Change linter from custom config to `purple_lints`.
+- [BREAKING] Rename `memoized` to `memoize()`.
+- Expose function typedefs.
+
+### Added
+
+- Add basic functions – `id`, `negate` and others.
+- Document everything.
+- `flip` extension.
+- `curry` extension.
+- `uncurry` extension.
+- `constant` extension.
+- `nullable` extension.
 
 ## [0.1.1] - 19.09.2021
 
 - Hotfix.
+
+## [0.1.0] - 19.09.2021
+
+- Initial version.
+